Create  Minimum_Knight_Moves.java
This commit includes a Java implementation for the Minimum Knight Moves problem on a chessboard of size N x N. The solution utilizes breadth-first search and considers the chessboard size as |x| + 2 * 301 and |y| + 2 * 301 to handle negative coordinates. The code is accompanied by a documentation section outlining the problem statement. #hacktoberfest